+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 0.2.1 - 2017-04-08
+
 ### Added
 
 - FuncEval #7